class21/pyproject.toml
[project]
name = "class21"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "fastmcp>=3.0.2",
    "openai-agents>=0.10.2",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
]
